+++ src/Clean.Architecture.UseCases/Contributors/GetContributor/GetContributorQuery.cs
@@ -1,5 +1,5 @@
 ﻿using Clean.Architecture.Core.ContributorAggregate;
 
-namespace Clean.Architecture.UseCases.Contributors.Get;
+namespace Clean.Architecture.UseCases.Contributors.GetContributor;
 
 public record GetContributorQuery(ContributorId ContributorId) : IQuery<Result<ContributorDto>>;
